+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 Unreleased Changes
 ------------------
 
+3.6.2 (2022-06-16)
+------------------
+
+* Issue - Fix DynamoDB session store to work with Rails 7.
+* Issue - Allow for dynamic message group ids in FIFO Queues.
+
 3.6.1 (2021-06-08)
 ------------------